scripts/requirements-review.txt
# Documentation review tool dependencies
python-frontmatter>=1.0.0
markdown>=3.3.0
beautifulsoup4>=4.9.0
PyYAML>=5.4.0
python-slugify>=5.0.0
colorama>=0.4.4
tqdm>=4.60.0
requests>=2.25.1
markdown-include>=0.6.0
